+++ src/behaviors.js
@@ -8,12 +8,99 @@
 // **MUST** override the class level behaviorsLookup
 // method for things to work properly.
 
-Marionette.Behaviors = (function(Marionette, _) {
-  // Borrow event splitter from Backbone
-  var delegateEventSplitter = /^(\S+)\s*(.*)$/;
+// Borrow event splitter from Backbone
+var delegateEventSplitter = /^(\S+)\s*(.*)$/;
 
-  function Behaviors(view, behaviors) {
+// Class to build handlers for `triggers` on behaviors
+// for views
+var BehaviorTriggersBuilder = Marionette.Class.extend({
+  constructor: function(view, behaviors) {
+    this._view      = view;
+    this._viewUI    = _.result(view, 'ui');
+    this._behaviors = behaviors;
+    this._triggers  = {};
+  },
 
+  // Main method to build the triggers hash with event keys and handlers
+  buildBehaviorTriggers: function() {
+    _.each(this._behaviors, this._buildTriggerHandlersForBehavior, this);
+    return this._triggers;
+  },
+
+  // Internal method to build all trigger handlers for a given behavior
+  _buildTriggerHandlersForBehavior: function(behavior, i) {
+    var ui = _.extend({}, this._viewUI, _.result(behavior, 'ui'));
+    var triggersHash = _.clone(_.result(behavior, 'triggers')) || {};
+
+    triggersHash = Marionette.normalizeUIKeys(triggersHash, ui);
+
+    _.each(triggersHash, _.bind(this._setHandlerForBehavior, this, behavior, i));
+  },
+
+  // Internal method to create and assign the trigger handler for a given
+  // behavior
+  _setHandlerForBehavior: function(behavior, i, eventName, trigger) {
+    // Unique identifier for the `this._triggers` hash
+    var triggerKey = trigger.replace(/^\S+/, function(triggerName) {
+      return triggerName + '.' + 'behaviortriggers' + i;
+    });
+
+    this._triggers[triggerKey] = this._view._buildViewTrigger(eventName);
+  }
+});
+
+var methods = {
+  behaviorTriggers: function(behaviorTriggers, behaviors) {
+    var triggerBuilder = new BehaviorTriggersBuilder(this, behaviors);
+    return triggerBuilder.buildBehaviorTriggers();
+  },
+
+  behaviorEvents: function(behaviorEvents, behaviors) {
+    var _behaviorsEvents = {};
+    var viewUI = this._uiBindings || _.result(this, 'ui');
+
+    _.each(behaviors, function(b, i) {
+      var _events = {};
+      var behaviorEvents = _.clone(_.result(b, 'events')) || {};
+      var behaviorUI = b._uiBindings || _.result(b, 'ui');
+
+      // Construct an internal UI hash first using
+      // the views UI hash and then the behaviors UI hash.
+      // This allows the user to use UI hash elements
+      // defined in the parent view as well as those
+      // defined in the given behavior.
+      var ui = _.extend({}, viewUI, behaviorUI);
+
+      // Normalize behavior events hash to allow
+      // a user to use the @ui. syntax.
+      behaviorEvents = Marionette.normalizeUIKeys(behaviorEvents, ui);
+
+      var j = 0;
+      _.each(behaviorEvents, function(behaviour, key) {
+        var match     = key.match(delegateEventSplitter);
+
+        // Set event name to be namespaced using the view cid,
+        // the behavior index, and the behavior event index
+        // to generate a non colliding event namespace
+        // http://api.jquery.com/event.namespace/
+        var eventName = match[1] + '.' + [this.cid, i, j++, ' '].join(''),
+            selector  = match[2];
+
+        var eventKey  = eventName + selector;
+        var handler   = _.isFunction(behaviour) ? behaviour : b[behaviour];
+
+        _events[eventKey] = _.bind(handler, b);
+      }, this);
+
+      _behaviorsEvents = _.extend(_behaviorsEvents, _events);
+    }, this);
+
+    return _behaviorsEvents;
+  }
+};
+
+var Behaviors = Marionette.Behaviors = Marionette.Class.extend({
+  constructor: function(view, behaviors) {
     if (!_.isObject(view.behaviors)) {
       return {};
     }
@@ -28,152 +115,58 @@ Marionette.Behaviors = (function(Marionette, _) {
     Behaviors.wrap(view, behaviors, _.keys(methods));
     return behaviors;
   }
+}, {
+  // Placeholder method to be extended by the user.
+  // The method should define the object that stores the behaviors.
+  // i.e.
+  //
+  // ```js
+  // Marionette.Behaviors.behaviorsLookup: function() {
+  //   return App.Behaviors
+  // }
+  // ```
+  behaviorsLookup: function() {
+    throw new Marionette.Error({
+      message: 'You must define where your behaviors are stored.',
+      url: 'marionette.behaviors.html#behaviorslookup'
+    });
+  },
 
-  var methods = {
-    behaviorTriggers: function(behaviorTriggers, behaviors) {
-      var triggerBuilder = new BehaviorTriggersBuilder(this, behaviors);
-      return triggerBuilder.buildBehaviorTriggers();
-    },
-
-    behaviorEvents: function(behaviorEvents, behaviors) {
-      var _behaviorsEvents = {};
-      var viewUI = this._uiBindings || _.result(this, 'ui');
-
-      _.each(behaviors, function(b, i) {
-        var _events = {};
-        var behaviorEvents = _.clone(_.result(b, 'events')) || {};
-        var behaviorUI = b._uiBindings || _.result(b, 'ui');
-
-        // Construct an internal UI hash first using
-        // the views UI hash and then the behaviors UI hash.
-        // This allows the user to use UI hash elements
-        // defined in the parent view as well as those
-        // defined in the given behavior.
-        var ui = _.extend({}, viewUI, behaviorUI);
-
-        // Normalize behavior events hash to allow
-        // a user to use the @ui. syntax.
-        behaviorEvents = Marionette.normalizeUIKeys(behaviorEvents, ui);
-
-        var j = 0;
-        _.each(behaviorEvents, function(behaviour, key) {
-          var match     = key.match(delegateEventSplitter);
-
-          // Set event name to be namespaced using the view cid,
-          // the behavior index, and the behavior event index
-          // to generate a non colliding event namespace
-          // http://api.jquery.com/event.namespace/
-          var eventName = match[1] + '.' + [this.cid, i, j++, ' '].join(''),
-              selector  = match[2];
-
-          var eventKey  = eventName + selector;
-          var handler   = _.isFunction(behaviour) ? behaviour : b[behaviour];
-
-          _events[eventKey] = _.bind(handler, b);
-        }, this);
-
-        _behaviorsEvents = _.extend(_behaviorsEvents, _events);
-      }, this);
-
-      return _behaviorsEvents;
+  // Takes care of getting the behavior class
+  // given options and a key.
+  // If a user passes in options.behaviorClass
+  // default to using that. Otherwise delegate
+  // the lookup to the users `behaviorsLookup` implementation.
+  getBehaviorClass: function(options, key) {
+    if (options.behaviorClass) {
+      return options.behaviorClass;
     }
-  };
 
-  _.extend(Behaviors, {
+    // Get behavior class can be either a flat object or a method
+    return Marionette._getValue(Behaviors.behaviorsLookup, this, [options, key])[key];
+  },
 
-    // Placeholder method to be extended by the user.
-    // The method should define the object that stores the behaviors.
-    // i.e.
-    //
-    // ```js
-    // Marionette.Behaviors.behaviorsLookup: function() {
-    //   return App.Behaviors
-    // }
-    // ```
-    behaviorsLookup: function() {
-      throw new Marionette.Error({
-        message: 'You must define where your behaviors are stored.',
-        url: 'marionette.behaviors.html#behaviorslookup'
-      });
-    },
+  // Iterate over the behaviors object, for each behavior
+  // instantiate it and get its grouped behaviors.
+  parseBehaviors: function(view, behaviors) {
+    return _.chain(behaviors).map(function(options, key) {
+      var BehaviorClass = Behaviors.getBehaviorClass(options, key);
 
-    // Takes care of getting the behavior class
-    // given options and a key.
-    // If a user passes in options.behaviorClass
-    // default to using that. Otherwise delegate
-    // the lookup to the users `behaviorsLookup` implementation.
-    getBehaviorClass: function(options, key) {
-      if (options.behaviorClass) {
-        return options.behaviorClass;
-      }
+      var behavior = new BehaviorClass(options, view);
+      var nestedBehaviors = Behaviors.parseBehaviors(view, _.result(behavior, 'behaviors'));
 
-      // Get behavior class can be either a flat object or a method
-      return Marionette._getValue(Behaviors.behaviorsLookup, this, [options, key])[key];
-    },
+      return [behavior].concat(nestedBehaviors);
+    }).flatten().value();
+  },
 
-    // Iterate over the behaviors object, for each behavior
-    // instantiate it and get its grouped behaviors.
-    parseBehaviors: function(view, behaviors) {
-      return _.chain(behaviors).map(function(options, key) {
-        var BehaviorClass = Behaviors.getBehaviorClass(options, key);
-
-        var behavior = new BehaviorClass(options, view);
-        var nestedBehaviors = Behaviors.parseBehaviors(view, _.result(behavior, 'behaviors'));
-
-        return [behavior].concat(nestedBehaviors);
-      }).flatten().value();
-    },
-
-    // Wrap view internal methods so that they delegate to behaviors. For example,
-    // `onDestroy` should trigger destroy on all of the behaviors and then destroy itself.
-    // i.e.
-    //
-    // `view.delegateEvents = _.partial(methods.delegateEvents, view.delegateEvents, behaviors);`
-    wrap: function(view, behaviors, methodNames) {
-      _.each(methodNames, function(methodName) {
-        view[methodName] = _.partial(methods[methodName], view[methodName], behaviors);
-      });
-    }
-  });
-
-  // Class to build handlers for `triggers` on behaviors
-  // for views
-  function BehaviorTriggersBuilder(view, behaviors) {
-    this._view      = view;
-    this._viewUI    = _.result(view, 'ui');
-    this._behaviors = behaviors;
-    this._triggers  = {};
+  // Wrap view internal methods so that they delegate to behaviors. For example,
+  // `onDestroy` should trigger destroy on all of the behaviors and then destroy itself.
+  // i.e.
+  //
+  // `view.delegateEvents = _.partial(methods.delegateEvents, view.delegateEvents, behaviors);`
+  wrap: function(view, behaviors, methodNames) {
+    _.each(methodNames, function(methodName) {
+      view[methodName] = _.partial(methods[methodName], view[methodName], behaviors);
+    });
   }
-
-  _.extend(BehaviorTriggersBuilder.prototype, {
-    // Main method to build the triggers hash with event keys and handlers
-    buildBehaviorTriggers: function() {
-      _.each(this._behaviors, this._buildTriggerHandlersForBehavior, this);
-      return this._triggers;
-    },
-
-    // Internal method to build all trigger handlers for a given behavior
-    _buildTriggerHandlersForBehavior: function(behavior, i) {
-      var ui = _.extend({}, this._viewUI, _.result(behavior, 'ui'));
-      var triggersHash = _.clone(_.result(behavior, 'triggers')) || {};
-
-      triggersHash = Marionette.normalizeUIKeys(triggersHash, ui);
-
-      _.each(triggersHash, _.bind(this._setHandlerForBehavior, this, behavior, i));
-    },
-
-    // Internal method to create and assign the trigger handler for a given
-    // behavior
-    _setHandlerForBehavior: function(behavior, i, eventName, trigger) {
-      // Unique identifier for the `this._triggers` hash
-      var triggerKey = trigger.replace(/^\S+/, function(triggerName) {
-        return triggerName + '.' + 'behaviortriggers' + i;
-      });
-
-      this._triggers[triggerKey] = this._view._buildViewTrigger(eventName);
-    }
-  });
-
-  return Behaviors;
-
-})(Marionette, _);
+});
