+++ src/controller.js
@@ -4,21 +4,7 @@
 // A multi-purpose object to use as a controller for
 // modules and routers, and as a mediator for workflow
 // and coordination of other objects, views, and more.
-Marionette.Controller = function(options) {
-  this.options = options || {};
-
-  if (_.isFunction(this.initialize)) {
-    this.initialize(this.options);
-  }
-};
-
-Marionette.Controller.extend = Marionette.extend;
-
-// Controller Methods
-// --------------
-
-// Ensure it can trigger events with Backbone.Events
-_.extend(Marionette.Controller.prototype, Backbone.Events, {
+Marionette.Controller = Marionette.Class.extend({
   destroy: function() {
     Marionette._triggerMethod(this, 'before:destroy', arguments);
     Marionette._triggerMethod(this, 'destroy', arguments);
@@ -26,13 +12,5 @@ _.extend(Marionette.Controller.prototype, Backbone.Events, {
     this.stopListening();
     this.off();
     return this;
-  },
-
-  // import the `triggerMethod` to trigger events with corresponding
-  // methods if the method exists
-  triggerMethod: Marionette.triggerMethod,
-
-  // Proxy `getOption` to enable getting options from this or this.options by name.
-  getOption: Marionette.proxyGetOption
-
+  }
 });
