+++ src/region.js
@@ -6,7 +6,7 @@
 // Manage the visual regions of your composite application. See
 // http://lostechies.com/derickbailey/2011/12/12/composite-js-apps-regions-and-region-managers/
 
-Marionette.Region = Marionette.Object.extend({
+Marionette.Region = Marionette.Class.extend({
   constructor: function (options) {
 
     // set options temporarily so that we can get `el`.
