+++ test/unit/setup/node.js
@@ -32,17 +32,17 @@ global.$ = global.jQuery = require('jquery');
 global._ = require('underscore');
 global.Backbone = require('backbone');
 global.Backbone.$ = global.$;
+global.Metal = require('backbone-metal');
 global.Marionette = Backbone.Marionette = {};
 require('backbone.wreqr');
 require('backbone.babysitter');
 global.slice = Array.prototype.slice;
 Marionette.Deferred = global.Backbone.$.Deferred;
-requireHelper('bind-entity-events');
+requireHelper('metal');
+requireHelper('utils');
 requireHelper('callbacks');
-requireHelper('trigger-method');
 requireHelper('helpers');
 requireHelper('dom-refresh');
-requireHelper('object');
 requireHelper('controller');
 requireHelper('app-router');
 requireHelper('application');
@@ -57,4 +57,3 @@ requireHelper('behavior');
 requireHelper('behaviors');
 requireHelper('region');
 requireHelper('region-manager');
-requireHelper('error');
